# comment_service/storage/repo/models.go
package repo

type Comment struct {
	Id           string
	PostId       string
	PostTitle    string
	UserId       string
	UserName     string
	PostUserName string
	Text         string
	CreatedAt    string
}